# Inline MemTable::KeyComparator