--- a/db/memtable.cc
+++ b/db/memtable.cc
@@ -57,6 +57,28 @@ MemTableOptions::MemTableOptions(const ImmutableCFOptions& ioptions,
       merge_operator(ioptions.merge_operator),
       info_log(ioptions.info_log) {}
 
+MemTableRep::KeyComparator::KeyComparator(const InternalKeyComparator& cmp)
+    : comparator_(cmp) {}
+
+int MemTableRep::KeyComparator::operator()(const char* prefix_len_key1,
+                                           const char* prefix_len_key2) const {
+  // Internal keys are encoded as length-prefixed strings.
+  Slice k1 = GetLengthPrefixedSlice(prefix_len_key1);
+  Slice k2 = GetLengthPrefixedSlice(prefix_len_key2);
+  return comparator_.Compare(k1, k2);
+}
+
+int MemTableRep::KeyComparator::operator()(const char* prefix_len_key,
+                                           const Slice& key) const {
+  // Internal keys are encoded as length-prefixed strings.
+  Slice a = GetLengthPrefixedSlice(prefix_len_key);
+  return comparator_.Compare(a, key);
+}
+
+const InternalKeyComparator& MemTableRep::KeyComparator::comparator() const {
+  return comparator_;
+}
+
 MemTable::MemTable(const InternalKeyComparator& cmp,
                    const ImmutableCFOptions& ioptions,
                    const MutableCFOptions& mutable_cf_options,
@@ -196,22 +218,6 @@ void MemTable::UpdateFlushState() {
   }
 }
 
-int MemTable::KeyComparator::operator()(const char* prefix_len_key1,
-                                        const char* prefix_len_key2) const {
-  // Internal keys are encoded as length-prefixed strings.
-  Slice k1 = GetLengthPrefixedSlice(prefix_len_key1);
-  Slice k2 = GetLengthPrefixedSlice(prefix_len_key2);
-  return comparator.Compare(k1, k2);
-}
-
-int MemTable::KeyComparator::operator()(const char* prefix_len_key,
-                                        const Slice& key)
-    const {
-  // Internal keys are encoded as length-prefixed strings.
-  Slice a = GetLengthPrefixedSlice(prefix_len_key);
-  return comparator.Compare(a, key);
-}
-
 Slice MemTableRep::UserKey(const char* key) const {
   Slice slice = GetLengthPrefixedSlice(key);
   return Slice(slice.data(), slice.size() - 8);
@@ -309,7 +315,7 @@ class MemTableIterator : public InternalIterator {
     if (!Valid()) {
       SeekToLast();
     }
-    while (Valid() && comparator_.comparator.Compare(k, key()) < 0) {
+    while (Valid() && comparator_.comparator().Compare(k, key()) < 0) {
       Prev();
     }
   }
@@ -358,7 +364,7 @@ class MemTableIterator : public InternalIterator {
  private:
   DynamicBloom* bloom_;
   const SliceTransform* const prefix_extractor_;
-  const MemTable::KeyComparator comparator_;
+  const MemTableRep::KeyComparator comparator_;
   MemTableRep::Iterator* iter_;
   bool valid_;
   bool arena_mode_;
@@ -710,7 +716,7 @@ void MemTable::Update(SequenceNumber seq,
     const char* entry = iter->key();
     uint32_t key_length = 0;
     const char* key_ptr = GetVarint32Ptr(entry, entry + 5, &key_length);
-    if (comparator_.comparator.user_comparator()->Equal(
+    if (comparator_.comparator().user_comparator()->Equal(
             Slice(key_ptr, key_length - 8), lkey.user_key())) {
       // Correct user key
       const uint64_t tag = DecodeFixed64(key_ptr + key_length - 8);
@@ -764,7 +770,7 @@ bool MemTable::UpdateCallback(SequenceNumber seq,
     const char* entry = iter->key();
     uint32_t key_length = 0;
     const char* key_ptr = GetVarint32Ptr(entry, entry + 5, &key_length);
-    if (comparator_.comparator.user_comparator()->Equal(
+    if (comparator_.comparator().user_comparator()->Equal(
             Slice(key_ptr, key_length - 8), lkey.user_key())) {
       // Correct user key
       const uint64_t tag = DecodeFixed64(key_ptr + key_length - 8);
@@ -835,7 +841,7 @@ size_t MemTable::CountSuccessiveMergeEntries(const LookupKey& key) {
     const char* entry = iter->key();
     uint32_t key_length = 0;
     const char* iter_key_ptr = GetVarint32Ptr(entry, entry + 5, &key_length);
-    if (!comparator_.comparator.user_comparator()->Equal(
+    if (!comparator_.comparator().user_comparator()->Equal(
             Slice(iter_key_ptr, key_length - 8), key.user_key())) {
       break;
     }