--- a/port/port_posix.h
+++ b/port/port_posix.h
@@ -84,7 +84,7 @@ const uint64_t kMaxUint64 = std::numeric_limits<uint64_t>::max();
 const int64_t kMaxInt64 = std::numeric_limits<int64_t>::max();
 const size_t kMaxSizet = std::numeric_limits<size_t>::max();
 
-static const bool kLittleEndian = PLATFORM_IS_LITTLE_ENDIAN;
+extern const bool kLittleEndian;
 #undef PLATFORM_IS_LITTLE_ENDIAN
 
 class CondVar;