--- a/util/coding.h
+++ b/util/coding.h
@@ -18,6 +18,7 @@
 #include <string.h>
 #include <string>
 
+#include "rocksdb/coding.h"
 #include "rocksdb/write_batch.h"
 #include "port/port.h"
 
@@ -105,23 +106,6 @@ inline uint64_t DecodeFixed64(const char* ptr) {
     uint64_t hi = DecodeFixed32(ptr + 4);
     return (hi << 32) | lo;
   }
-}
-
-// Internal routine for use by fallback path of GetVarint32Ptr
-extern const char* GetVarint32PtrFallback(const char* p,
-                                          const char* limit,
-                                          uint32_t* value);
-inline const char* GetVarint32Ptr(const char* p,
-                                  const char* limit,
-                                  uint32_t* value) {
-  if (p < limit) {
-    uint32_t result = *(reinterpret_cast<const unsigned char*>(p));
-    if ((result & 128) == 0) {
-      *value = result;
-      return p + 1;
-    }
-  }
-  return GetVarint32PtrFallback(p, limit, value);
 }
 
 // -- Implementation of the functions declared above
@@ -315,14 +299,6 @@ inline bool GetLengthPrefixedSlice(Slice* input, Slice* result) {
   } else {
     return false;
   }
-}
-
-inline Slice GetLengthPrefixedSlice(const char* data) {
-  uint32_t len = 0;
-  // +5: we assume "data" is not corrupted
-  // unsigned char is 7 bits, uint32_t is 32 bits, need 5 unsigned char
-  auto p = GetVarint32Ptr(data, data + 5 /* limit */, &len);
-  return Slice(p, len);
 }
 
 inline Slice GetSliceUntil(Slice* slice, char delimiter) {