The current call stack when InlinkeSkipList::Insert is called will involve:
(0) BytewiseComparator::Compare()
(1) InternalKeyComparator::Compare()
(2) MemTable::KeyComparator::operator()
(3) InlineSkipList::KeyIsAfterNode()
(4) InlineSkipList::Insert()

(0), (1) and (2) are all virtual methods so they cannot be inlined by compiler. Since the whole key comparison logic is relatively simple, the overhead of function call is not very small. This PR tries to make (1) (2) (3) inline by:
* Remove MemTable::KeyComparator. Move its logic to its parent MemTableRep::KeyComparator and put the definition in public header include/rocksdb/memtablerep.h.
* Move InternalKeyComparator to public header include/rocksdb/internal_comparator.h.
* Move helper function needed by MemTable::KeyComparator and InternalKeyComparator to public header include/rocksdb/coding.h.
* Add "inline" keyword to InlineSkipList::KeyIsAfterNode() to suggest compiler inline it.

To compare the performance with master, I'm running the following command:
./db_bench --benchmarks=filluniquerandomdeterministic --disable_auto_compactions=true --num_levels=1 --disable_wal --write_buffer_size=2000000000 --num=10000000

master (3 times):
filluniquerandomdeterministic :       5.173 micros/op 193322 ops/sec;   21.4 MB/s
filluniquerandomdeterministic :       5.221 micros/op 191532 ops/sec;   21.2 MB/s
filluniquerandomdeterministic :       5.092 micros/op 196394 ops/sec;   21.7 MB/s

with the PR (3 times):
filluniquerandomdeterministic : 4.987 micros/op 200522 ops/sec; 22.2 MB/s
filluniquerandomdeterministic : 5.060 micros/op 197627 ops/sec; 21.9 MB/s
filluniquerandomdeterministic : 5.051 micros/op 197984 ops/sec; 21.9 MB/s

It shows roughly 5% more throughput.

`perf stat` command result:
master: https://gist.github.com/yiwu-arbug/a110a2b230aba171c3f4f608f1f5a38f
with the PR: https://gist.github.com/yiwu-arbug/f37f8e78ee81277d66492dd0dbbe37be

It shows a 10% reduction in number of CPU instructions executed.