@siying 2017-02-27
We are moving lots of internal code under include/rocksdb. If we move KeyComparator out of MemTableRep and forward declare it, can we avoid those public implementation?