@yiwu-arbug 2017-02-27
This patch is trying to make as much of the logic inlined under -O2 since we are using -O2 in Makefile. Just check we are using -O3 in fbcode. Shall we update our Makefile to have -O3 for non-debug build?