@siying 2017-02-27
I remembered that there was some concerned about using -O3 but I forgot. Are you saying, without explicit inlining, we don't see the benefits in your benchmark?