@yiwu-arbug 2017-02-27
right. I have to force the compiler to inline those functions to see the benefits, but that's under -O2. With -O3 compiler will inline more aggressively so I have to redo my test to see.