@siying 2017-02-27
@yiwu-arbug sure let's try -O3.