@gfosco 2018-01-10
Closing this via automation due to lack of activity.  If discussion is still needed here, please re-open or create a new/updated issue.